requirements.txt
streamlit>=1.29.0
supabase>=2.3.0
python-dotenv>=1.0.0
pandas>=2.2.0
plotly>=5.18.0
openpyxl>=3.1.2
reportlab>=4.0.7
requests>=2.31.0
Pillow>=10.1.0
python-dateutil>=2.8.2
pytz>=2023.3
